refactor(AnalysisRoutes): move componet to its own directory
