+++ src/components/router/Router.tsx
@@ -5,7 +5,7 @@ import Login from '~/pages/login/Login'
 import ValidateAuthentication from '~components/auth/Authentication'
 import Home from '~/pages/home/Home'
 import AuthenticatedLayout from '~/components/layouts/AuthenticatedLayout/AuthenticatedLayout'
-import { analysisRoutes } from './analysisRoutes'
+import { analysisRoutes } from './AnalysisRoutes/AnalysisRoutes'
 import { parametersRoutes } from './parametersRoutes'
 import { reportsRoutes } from './reportsRoutes'
 import { transactionsRoute } from './transactionsRoutes'
